Move webpack.base.config.js into webpack.config.js So main config is all in one place, makes it easier to search for and reason about.
